refactor: (GenAI) Reorganized Function calling Samples (Group A)  (#12591)
* New files for Function Calling folder

* Marks old files to be deleted

* Renamed tests

* Fix comments

* Update comments

* Rewrite tests

* Changed comments -> JSON schema format
