+++ generative_ai/function_calling/examples_function_calling_test.py
@@ -11,15 +11,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-# TODO: Delete this file after approval /function_calling/examples_function_calling_test.py
+# TODO: Rename this file after deleting /generative_ai/function_calling_test.py
+import advanced_example
 
 import backoff
 
+import basic_example
+
 from google.api_core.exceptions import ResourceExhausted
-
-import pytest
-
-import function_calling
 
 
 summary_expected = [
@@ -38,15 +37,15 @@ response_expected = [
 
 @backoff.on_exception(backoff.expo, ResourceExhausted, max_time=10)
 def test_function_calling() -> None:
-    response = function_calling.generate_function_call()
+    response = basic_example.generate_function_call()
     assert all(x in str(response.text) for x in summary_expected)
     assert all(x in str(response) for x in response_expected)
 
 
-# TODO: Remove skip once b/336973838 is resolved.
-@pytest.mark.skip("Service currently returning INVALID_ARGUMENT")
 @backoff.on_exception(backoff.expo, ResourceExhausted, max_time=10)
-def test_function_calling_advanced() -> None:
-    response = function_calling.generate_function_call_advanced()
-    assert all(x in str(response.text) for x in summary_expected)
-    assert all(x in str(response) for x in response_expected)
+def test_function_calling_advanced_function_selection() -> None:
+    response = advanced_example.generate_function_call_advanced()
+    assert (
+        "Pixel 8 Pro 128GB"
+        in response.candidates[0].function_calls[0].args["product_name"]
+    )
